Opesci: Separating StaggeredGrid, Variable, Fields and CodePrinter classes
