+++ opesci/__init__.py
@@ -1,2 +1,4 @@
-import grid
+from variable import *
+from fields import *
+from staggeredgrid import *
 import grid_test
